TownyExplodeEvent commit.
- Adds TownyExplodeEvent(Location location, boolean canExplode)
- Used via the TownyActionEventExecutor in the Listener classes to
decide the outcome of an explosion.
- Deletes the ExplosionUtil.
- Separates the war explosion tests, moving them into the
WarZoneListener, where the TownyExplodeEvent outcome is altered if
needed.


TODO: Work out some way of allowing for Block Damage and Entity Damage to be separated from each other so that wars can have TNT that hurts entities but does not destroy the terrain. Potentially two events needed.